Added CharacterInstance ScriptableObject, which holds a character, plus their actions in a specific scene.
Removed Character field from DialogueLine, as the speaker is now the
first character in the list of CharacterInstances.

Added CharacterPortrait class, which is attached to a CharacterPortait
UI object, and listens for scene updates from
CharacterPortraitController, and hides or shows itself accordingly.

Added CharacterPortraitController, which is a state manager singleton
that oversees all of the character portraits.

DialogueController.ReadDialogue now shows character portraits.

Updates NameDisplayController to use CharacterInstances instead of a
Character.

Added filename variables to SaveController.